requirements.txt
apache-libcloud==2.4.0
Flask==1.0.2
Flask-Cors==3.0.6
Flask-HTTPAuth==3.2.4
gevent==1.3.5
gunicorn==19.9.0
mongoengine==0.15.0
pymongo==3.6.1
python-faasclient==0.5.3
cachetools==4.1.1
kubernetes==11.0.0
oauthlib==3.1.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.1
requests-oauthlib==1.3.0
rsa==4.6
websocket-client==0.57.0
google-api-core==1.24.1
google-api-python-client==1.12.8
google-auth==1.24.0
google-auth-httplib2==0.0.4
googleapis-common-protos==1.52.0